Extract shared chat prompt core module
